requirements.txt
fastapi==0.115.4
motor==3.6.0
pydantic==2.9.2
python-dotenv==1.0.1
starlette==0.41.2
uvicorn==0.32.0
